[KAN-5] Frontend: Create a page for user login and user creation
