+++ pom.xml
@@ -19,6 +19,12 @@
             <artifactId>json</artifactId>
             <version>20230227</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+            <version>4.5</version>
+       </dependency>
     </dependencies>
 
     <build>
