+++ src/main/resources/html/login.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+
+<html>
+<head>
+  <title></title>
+  <!-- <link rel="stylesheet" href="src/main/css/VT-Fantasy/styles.css"> -->
+</head>
+
+<body>
+  <div>
+    <br><hr><br>
+    <h1>Login:</h1>
+    <form method="post" action="/verify_login">
+      <label for="email">Email:</label><br>
+      <input type="text" id="email" name="name"><br>
+      <label for="password">Password:</label><br>
+      <input type="password" id="password" name="password"><br><br>
+      <input type="submit" value="Login">
+    </form>
+    <br><hr><br>
+    <form method="post" action="/register">
+      <label for="fname">First Name:</label><br>
+      <input type="text" id="fname" name="fname"><br>
+      <label for="lname">Last Name:</label><br>
+      <input type="text" id="lname" name="lname"><br>
+      <label for="email">Email:</label><br>
+      <input type="text" id="email" name="email"><br>
+      <label for="password">Password:</label><br>
+      <input type="password" id="password" name="password"><br>
+      <label for="verify_password">Verify Password:</label><br>
+      <input type="password" id="verify_password" name="verify_password"><br><br>
+      <input type="submit" value="Register">
+    </form>
+  <div>
+</body>
+
+</html>
